+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](https://semver.org/).
 
-## 1.5.1
+## 1.6.0 (14. February 2024)
 
 + [#19](https://github.com/luyadev/yii-helpers/pull/19) Fixed issue with ordinal numbers.
++ [#21](https://github.com/luyadev/yii-helpers/pull/21) Added `ImportHelper::csvFromResource()` method to import CSV from a resource object like `fopen()`.
 
 ## 1.5.0 (26. October 2023)